image-search/requirements.txt
fastapi
uvicorn[standard]
pydantic
torch
transformers
Pillow
numpy
